pyproject.toml
[project]
name = "kingdom-archives-scraper"
version = "0.1.0"
description = "Scraper for kingdomarchives.com assets with descriptive organization"
authors = [{name = "Client"}]
requires-python = ">=3.10"
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "tqdm>=4.66.0",
    "tenacity>=8.2.0",
    "python-slugify>=8.0.0"
]

[project.scripts]
kingdom-archives-scraper = "kingdom_archives.cli:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
